# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JEE Backlog Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /history/undo:
    post:
      operationId: undoLastAction
      tags: [history]
      summary: Undo the last action
      responses:
        "200":
          description: Undo successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UndoResult"
        "400":
          description: Nothing to undo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Chapter:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        name:
          type: string
        totalLectures:
          type: integer
        completedLectures:
          type: integer
        remainingLectures:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - subjectId
        - name
        - totalLectures
        - completedLectures
        - remainingLectures
        - createdAt
        - updatedAt
...
    UndoResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        chapter:
          $ref: "#/components/schemas/Chapter"
      required:
        - success
        - message
        - chapter
...