# lib/api-zod/src/generated/types/undoResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * JEE Backlog Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Chapter } from './chapter';

export interface UndoResult {
  success: boolean;
  message: string;
  chapter: Chapter;
}